Create .pylintrc and modify some files to let the pylint work.
